% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{auroc}
\alias{auroc}
\title{calculate area under ROC curve}
\usage{
auroc(score, bool)
}
\arguments{
\item{score}{a score (here we use partial correlation values) for each edge}

\item{bool}{logical, indicating the existance of an edge in a network.}
}
\value{
the value of AUROC
}
\description{
calculate area under ROC curve
}
\references{
https://blog.mbq.me/augh-roc/
}
