% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{check_ident_list}
\alias{check_ident_list}
\title{check if the list of block dimensions are the same for the different conditions}
\usage{
check_ident_list(xlist)
}
\arguments{
\item{xlist}{a list of vectors}
}
\description{
check if the list of block dimensions are the same for the different conditions
}
