% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{CountMap}
\alias{CountMap}
\title{Map scRNA-seq counts to a known covariance structure}
\usage{
CountMap(sigma, ngene, n, a1 = 3, b1 = 1, a20 = 2, b20 = 3, a30 = 1, b30 = 10)
}
\arguments{
\item{sigma}{a covairance matrix}

\item{ngene}{number of genes}

\item{n}{number of samples}

\item{a20}{the hyperparameter for generating beta, where beta~Gamma(a20,b20)}

\item{b20}{the hyperparameter for generating beta, where beta~Gamma(a20,b20)}

\item{a10}{the hyperparameter for generating alpha, where alpha~Gamma(a10,b10)}

\item{b10}{the hyperparameter for generating alpha, where alpha~Gamma(a10,b10)}

\item{a3}{the hyperparameter for generating pij, where pij~beta(a3, b3)}

\item{b3}{the hyperparameter for generating pij, where pij~beta(a3, b3)}
}
\value{
a list of i) raw count data matrix; ii) adjacency matrix; iii) the true expression mean level theta matrix; iv) zij: the dropout matrix; v) sigma: the covariance matrix; vi) precision: the precision matrix.
}
\description{
Map scRNA-seq counts to a known covariance structure
}
