% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covcpd_functions.R
\name{covcpd}
\alias{covcpd}
\title{covcpd change point detection by covariance testing}
\usage{
covcpd(
  X,
  nperm = 100,
  siglvl = 0.05,
  k = 30,
  maxseg = NULL,
  verbose = T,
  search_by = NULL
)
}
\arguments{
\item{X}{raw count matrix, with genes by samples}

\item{nperm}{number of permutation}

\item{siglvl}{significance level, default is 0.05}

\item{k}{the minimum length of a segment / partition}

\item{maxseg}{maximum number of segments / partitions}

\item{verbose}{whether to print out the processing info while runing the funtion.}

\item{search_by}{the space between candidate change points.}
}
\value{
cps: change points;
cps_order: change points by the order when they are found;
candtau: candidate change points that were tested;
pvals: p-values for the candidate change points;
pthresh: the learned p-value threshold.
}
\description{
covcpd change point detection by covariance testing
}
