% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generateBlki}
\alias{generateBlki}
\title{generate the small matrix in block i}
\usage{
generateBlki(ni, ud = c(-100:-60, 60:100)/100)
}
\arguments{
\item{ni}{the number of rows and columns in block i}

\item{ud}{the range of uniform distribution that to generate the numbers in the block}
}
\description{
generate the small matrix in block i
}
