% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generateSigmaList}
\alias{generateSigmaList}
\title{generate a list of covariance matrices with designed structures}
\usage{
generateSigmaList(
  nivec.list,
  ud = c(-100:-60, 60:100)/100,
  structure = "Identical S, Identical W",
  diffblk = NULL
)
}
\arguments{
\item{nivec.list}{a list of vectors. Each vector specifies the dimensions of the blocks under a condition.
For example, c(20,20) means a covariance matrix is composed of two block matrices, each having 20 genes in the block.}

\item{ud}{The uniform distribution range. Default is c(-100:-60, 60:100)/100}

\item{structure}{structure S, Weight W: "Identical S, Identical W", "Identical S, Diff W", "Diff S, Identical W", "Diff S, Diff W".
If using the structure "Diff S, Identical W", setting some structure different while keep the rest of structure and weights the same, specify diffblk = list(diffblk1=1, diffblk2=c(1,2),...)}

\item{diffblk}{a list of vectors indicating the location of blocks that have different structure/weights.
For example, nivec.list = list(nivec= c(10,20,30,40, rep(20,5)),nivec2 = rep(20,10)); diffblk = list(1:4,1:5)}
}
\value{
a list of covariance matrices
}
\description{
generate a list of covariance matrices with designed structures
}
