% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{getGENEbox}
\alias{getGENEbox}
\title{plot: boxplot for a specific gene by group}
\usage{
getGENEbox(countmat, group, gene)
}
\arguments{
\item{countmat}{raw count matrix, gene by sample}

\item{group}{group identities of cells}

\item{gene}{the name of a TARGET GENE. e.g.: "MYC"}
}
\value{
A plot
}
\description{
plot: boxplot for a specific gene by group
}
