% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{getGENEhist}
\alias{getGENEhist}
\title{plot: histogram of raw counts by group}
\usage{
getGENEhist(countmat, group, gene, xlabel.pos = 500)
}
\arguments{
\item{countmat}{the raw count matrix , gene by sample}

\item{group}{a vector of group identities for cells/samples}

\item{gene}{the name of a TARGET GENE. e.g.: "MYC"}

\item{xlabel.pos}{the position of samplesize/expression rate information}
}
\value{
a plot
}
\description{
plot: histogram of raw counts by group
}
