% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{getObservedList}
\alias{getObservedList}
\title{create list of matrices according to conditions}
\usage{
getObservedList(mtx, group = NULL, min.cells = 10, geneSet = NULL)
}
\arguments{
\item{mtx}{raw count matrix, genes by samples}

\item{group}{vector of group identities. For example: c("A","A","B","C")}

\item{min.cells}{integer, keep genes with at least __ cells expressed. Default is 10.}

\item{geneSet}{a vector of genes of interest}
}
\value{
a list of matrices by group
}
\description{
create list of matrices according to conditions
}
