% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{getPathFreq}
\alias{getPathFreq}
\title{Given a set of genes, get the frequencies of genes in each pathway.}
\usage{
getPathFreq(meta, VARPATH = "SigmaMiniMap.Term")
}
\arguments{
\item{meta}{a data.frame including the pathway information of genes}

\item{VARPATH}{the variable name for pathways}
}
\value{
a table showing the frequencies of genes in each pathway
}
\description{
Given a set of genes, get the frequencies of genes in each pathway.
}
