% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuningParameterSelection.R
\name{huge_stars_joint}
\alias{huge_stars_joint}
\title{The StARS tuning parameter selection method from package 'huge'. Adapted to JGL models.}
\usage{
huge_stars_joint(
  inputlist,
  lambda = NULL,
  nlambda = 10,
  lambda.min.ratio = NULL,
  stars.subsample.ratio.vec = NULL,
  rep.num = 20,
  stars.thresh = 0.1,
  lam2init = NULL,
  lam1init = NULL,
  verbose = T,
  ...
)
}
\arguments{
\item{inputlist}{a list of matrices with Gaussian distributed counts}

\item{lambda}{a vector of candidate tuning parameter values. If NULL, the values will be generated according to the data.}

\item{nlambda}{interger, number of candidate lambda values}

\item{lambda.min.ratio}{numeric, the ratio of lambda.min / lambda.max, if NULL, it willbe set to 0.1}

\item{stars.subsample.ratio.vec}{vector, the subsample ratios. If NULL, it will be decided by data}

\item{rep.num}{integer, the number of subsampling}

\item{stars.thresh}{the threshold value in StARS}

\item{lam2init}{the initial value of lambda2 in JGL. This is set when try to optimize over lambda1}

\item{lam1init}{the initial value of lambda1 in JGL. This is set when try to optimize over lambda2}

\item{verbose}{logical, whether show processing messages.}
}
\value{
lambda: a vector of candidate lambda used. opt.lambda: the selected lambda value
}
\description{
The StARS tuning parameter selection method from package 'huge'. Adapted to JGL models.
}
