% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{Map2Pathways}
\alias{Map2Pathways}
\title{Map the gene set of interest to the known pathways}
\usage{
Map2Pathways(
  partcorr.list,
  conditions = NULL,
  GeneInterest = NULL,
  pathwayRef,
  pathwayRef_geneVariable,
  pathwayRef_pathVariable,
  threshold = 0
)
}
\arguments{
\item{partcorr.list}{a list of estimated partial correlation matrices}

\item{conditions}{a vector of the names of the corresponding partial correlation matrices}

\item{GeneInterest}{character. e.g. "MYC"}

\item{pathwayRef}{a data frame, including the gene and pathway information. e.g. column1 is genes, column2 is the pathways this gene belongs to according to Sigma.Mini.Map}

\item{pathwayRef_geneVariable}{character, the column name of gene variable}

\item{pathwayRef_pathVariable}{character, the column name of pathway variable}

\item{threshold}{the cutoff value for partial correlation, for visualization and GSEA purpose.}
}
\value{
x.connect: the list of genes that connected to the GeneInterest under each condition

GSEA.table: the final GSEA table
}
\description{
Map the gene set of interest to the known pathways
}
