% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JGNsc_functions.R
\name{mcImpute.R}
\alias{mcImpute.R}
\title{Matrix completion imputation method}
\source{
https://github.com/aanchalMongia/McImpute_scRNAseq
}
\usage{
mcImpute.R(
  data,
  preprocess = T,
  eps = 1e-12,
  normfac = 1,
  insweep = 20,
  tol = 1e-04,
  decfac = 0.7,
  min_count = 2,
  min_cells = 3,
  verbose = F
)
}
\arguments{
\item{data}{a matrix, samples by genes}

\item{preprocess}{logical, whether to preprocess data: Removing BAD genes, median normalization and log-transformation.}

\item{eps}{a small positive number that controls convergence, when calculating the Frobenius norm.}

\item{normfac}{normalize factor, default is 1.}

\item{insweep}{the max number of iterations when minimizing the rank of matrix given a tuning parameter(lambda) value}

\item{tol}{a small positive number that controls convergence}

\item{decfac}{decay factor, controls the rate when searching for the softthresholding parameter.}

\item{min_count}{for filtering step, minimum number of counts that a gene expressed in a cell}

\item{min_cells}{for filtering step, minimum number of cells that a gene expressed in.}

\item{verbose}{logical, show message.}
}
\value{
an imputed matrix
}
\description{
the original algorithm was implemented in Matlab. Here, we rewrite it in R.
}
\references{
https://www.frontiersin.org/articles/10.3389/fgene.2019.00009/full
}
