% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{plot_onenet}
\alias{plot_onenet}
\title{plot one network}
\usage{
plot_onenet(
  parcorr,
  mode = "circle",
  gname = "",
  lsize = 2.5,
  nodecolor = "grey",
  family.vec = 0,
  circlenet = F
)
}
\arguments{
\item{parcorr}{partial correlation matrix}

\item{mode}{ggnet2 mode, could be "circle" or "spring",...}

\item{gname}{character, graph name.}

\item{lsize}{numeric, label size}

\item{nodecolor}{vector/character, specify the color for each node}

\item{family.vec}{vector, assign nodes to families. In other words, it could be the classification / covariates of the nodes.}

\item{circlenet}{logical, whether plot the network in a circle shape.}
}
\value{
a network figure
}
\description{
plot one network
}
