% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{prec2partialcorr}
\alias{prec2partialcorr}
\title{Transform the precision matrix to partial correlation matrix}
\usage{
prec2partialcorr(mat)
}
\arguments{
\item{mat}{the precision matrix}
}
\value{
a partial correlation matrix
}
\description{
Transform the precision matrix to partial correlation matrix
}
