% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{PRROC_pr}
\alias{PRROC_pr}
\title{calculate area under precision-recall curve}
\usage{
PRROC_pr(z, trueadj.list)
}
\arguments{
\item{z}{a list of scores (here we use partial correlation values)}

\item{trueadj.list}{a list of true adjacency matrices}
}
\value{
AUPRC
}
\description{
calculate area under precision-recall curve
}
