% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JGNsc_functions.R
\name{RunJGNsc}
\alias{RunJGNsc}
\title{RunJGNsc framework main function}
\usage{
RunJGNsc(
  observed.list,
  warm = 1000,
  iter = 5000,
  mask.rate = 0.15,
  nrep = 50,
  min.cell = 3,
  runNetwork = F,
  l1.vec = NULL,
  l2.vec = NULL
)
}
\arguments{
\item{observed.list}{the list containing the matrices of K conditions. dim: genes by samples}

\item{warm}{the number of warm steps in the MCMC procedure}

\item{iter}{the number of iterations in the MCMC procedure}

\item{mask.rate}{iterative imputation procedure}

\item{nrep}{number of iterations in the imputation procedure, default is 50}

\item{min.cell}{the min number of cells that should a gene should express}

\item{runNetwork}{logical. Run the joint graphical lasso procedure or not.}

\item{l1.vec}{if runNetwork=T, the vector of candidate values for the tuning parameter lambda1}

\item{l2.vec}{if runNetwork=T, the vector of candidate values for the tuning parameter lambda2}
}
\value{
theta.star.npn: the imputed and Gaussian transformed list of matrices

if runNetwork = T, then it will return the JGL results (the precision matrices can be accessed by result$JGL$theta)
, aic.table (the AIC values with corresponding tuning parameter candidates), partcorr (the list of partial correlation matrices)
}
\description{
RunJGNsc framework main function
}
