% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JGNsc_functions.R
\name{softThreshold}
\alias{softThreshold}
\title{soft threshold function}
\usage{
softThreshold(x, thld = 0.5)
}
\arguments{
\item{x}{numeric}

\item{thld}{the threshold value}
}
\value{
a number by soft thresholding
}
\description{
soft threshold function
}
