% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuningParameterSelection.R
\name{stars.select}
\alias{stars.select}
\title{Tuning parameter selection in JGL by StARS}
\usage{
stars.select(inputlist, nlamb = 15, rep.num = 15, lambvec = NULL)
}
\arguments{
\item{inputlist}{a list of matrices with Gaussian distributed counts}

\item{nlamb}{integer, specify the number of tuning parameter candidate}

\item{rep.num}{integer, the number of subsampling to perform, default is 15}

\item{lambvec}{vector, the candidate tuning parameter values for both lambda1 and lambda2}
}
\value{
stars.l1: the stars tuning parameter selection result when fixing lambda2

stars.l2: the stars tuning parameter selection result when fixing lambda2

res: the JGL result ;

runtime: run time
}
\description{
Tuning parameter selection in JGL by StARS
}
