% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ThresholdPartCorr}
\alias{ThresholdPartCorr}
\title{Thresholding a Partial correlation matrix}
\usage{
ThresholdPartCorr(x, y)
}
\arguments{
\item{x}{Gaussian matrix with samples x genes}

\item{y}{the estimated partial correlation matrix}
}
\description{
Thresholding a Partial correlation matrix
}
