% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ThresholdPartCorrList}
\alias{ThresholdPartCorrList}
\title{Thresholding a list of Partial correlation matrices}
\usage{
ThresholdPartCorrList(x, y)
}
\arguments{
\item{x}{a list of Gaussian matrices with samples x genes}

\item{y}{a list of the estimated partial correlation matrices}
}
\description{
Thresholding a list of Partial correlation matrices
}
