% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{trunc_precision}
\alias{trunc_precision}
\title{truncate the (precision) matrix}
\usage{
trunc_precision(z, threshold = 0.01)
}
\arguments{
\item{z}{a matrix}

\item{threshold}{the threshold value, where if a value's absolute value is less than this, it will be truncated to zero.}
}
\value{
a truncated matrix
}
\description{
truncate the (precision) matrix
}
